Update patch set 1
Patch Set 1:

i think the only supported phone with associated NFC binaries is the Nexus S....however, if beam works on your phone, i could be wrong.  

the way to tell if the NFC stack is working on your phone with ICS, Google added a noise when the NFC intent fires (even if there are no apps using it).  so if you put your phone to the tag and it "tweets" you know the NFC stack is using the reader.  this happens on the Nexus S...(my test phone).   

for this patch, you aren't reading tags, really, you are reading RFID readers instead.   the NFC intent will fire on the first PCD command from the RFID reader to the phone.  the reader should be a Type A PCD reader if there is no app installed on the phone to accept the reader command (this will give you the tweet you are looking for).  because you never want the phone to be both type A and type B at the same time,  the app level controls a type B reader emulation state with the setForeground utility.

-------------
I know you only added functionality. My question was more general for the NFC support in the current CM9 source tree. It seems somewhat broken as I can read tags a friend with a stock build + appropriate app.

This patch seems fine just from looking at it but I can't try it because of the above issue with tag reading. I have tried both NFC TagInfo and TagInfo that work on the stock rom.

Patch-set: 1