Update patch set 1
Patch Set 1:

I find it very strange that the NFC code in AOSP wouldn't be compatible with the Galaxy Nexus (tuna and toro) since it's now their flagship device and it supports Google Wallet (can't try that either since it's not supported in Sweden). And no, I get no noise when trying to scan a tag.

Could someone with a Galaxy Nexus verify if NFC works correctly with AOSP-based builds? Just in-case it's just me doing something wrong.

Oh, I thought it added more tag support at first based on the commit message. Googling shows it refers to RFID as you mentioned.

Patch-set: 1