Update patch set 1
Patch Set 1:

OS Card Emulation is fully implemented currently.  works like a charm.  this fix helps to play nice with gwallet which uses the SE for emulation.  this flag:  
//#define CONFIG_CE_DEFAULT
(ln 54 of com_android_nfc_NativeNfcManager.cpp)
could actually be controlled with a checkbox in the wireless services menu under the checkbox for NFC.  it determines the default operating state of the NFC service as it relates to how card emulation is done (SE & OS), or when checked (just OS).  if it is just OS, the advantage to the user is making card emulation from the OS when the current app is not in the foreground.....just wave the phone no matter what app is up is possible. 

the other setting (SE & OS) makes the app the controller as it's in the foreground.

Patch-set: 1