Update patch set 1
Patch Set 1:

it's not firmware, i can confirm that.
confirmed from log, i'm running the same.  i think it is a supplemental SE issue.  i.e., it is physically possible to have multiple SE (one in a sim card also with the embedded SMX one) plugged into the device.  i think we are seeing a conflict in this case.  i have no hardware to confirm this and remedy this.  i'm working on a solution.
it is a very rare case at this point.  may do a short term work-around that disables the conflict (and HCE by default in the event of this case).  while looking for the long term solution.  not sure yet, still looking into it.

Patch-set: 1