Update patch set 2
Patch Set 2: Code-Review-1

Anyway, your description of the problem doesn't seem right. getPackagesForUid could return more than a pid for a uid, but not of different uids. Basically, you can get a pid from Launcher from the uid of Settings. In that case the uid should be of the Launcher.

Patch-set: 2