Update patch set 2
Patch Set 2:

> Anyway, your description of the problem doesn't seem right.
 > getPackagesForUid could return more than a pid for a uid, but not
 > of different uids. Basically, you can get a pid from Launcher from
 > the uid of Settings. In that case the uid should be of the
 > Launcher.

Yes, as you say, getPackagesForUid could return more than a pid for a uid, but we use the first one as the foreground activity package, this is not correct.
You can do some test, add some log in P2pLinkManager.java, you can see if we are in Settings, and value of the getPackagesForUid()[0] must not be Settings.
So i modify some code to search the package of current foreground pid of current uid.

Patch-set: 2