Update patch set 4
Patch Set 4: Code-Review-1

guotao, after debugging the code I can't reproduce the problem you described. The only way it can be wrong is that Launcher3 and Settings shared the same uid (1000), which is a security flaw. Uid 1000 is asigned to system which is shared by various packages, but not for the launcher, so basically, getPackagesForUid cannnot return the launcher package.
And reading the code again, foregroundUids is obtained from mForegroundUtils.getForegroundUids(), that is used by getForegroundActivityPkg to retrieve the current package. This can only works because you use getRunningAppProcesses() 
which can return the processes in a different order getPackagesForUid, and, luckily, you obtain the Launcher process before Settings, which is just a matter of lucky.

Patch-set: 4
Label: Code-Review=-1