Update patch set 4
Patch Set 4:

Could you debug which uid you get from foregroundUids.get(0) and the pkgs returned by getPackagesForUid() for this uid in P2pLinkManager when you are beaming in Launcher3? In my tests i can't reproduce your problem. When I'm beaming in Settings uid is 1000 (system) that efectively is shared by different processes (settings, telecomm, android, ...), but launcher3 is not present, and beaming in Launcher3 returns the uid of the Launcher3 (Trebuchet) app that doesn't have a sharedid but have two process (com.android.launcher3 and com.android.launcher3:wallpaper_chooser). I can only think in three different cases that could cause the behaviour you are seeing:

1) Settings and Launcher3 shares the same id. In that case getPackagesForUid() will return both packages and maybe it could be obtain the wrong package. But as i said this is a security flaw, since Launcher3 must not be running as system user. This could be a problem for the Settings app because there are other packages running at the same uid, but in the other packages rarelly have foreground activities, so it behaves correctly.
2) If the foreground activity is Launcher3 and foregroundUids.get(0) returns the uid of Launcher3, then getPackagesForUid() always must return the Launcher3 package. I can't see any problem here.
3) If the foreground activity is Launcher3 and foregroundUids.get(0) returns a different uid of Launcher3, then foregroundUids is out of sync. In that case, the new method to obtain the foreground package name is also out of sync, so we aren't going to obtain the right package. Because foregroundUids is a clone of mForegroundUidPids in ForegroundUtils. The new method will get the same uid, that contains only pids of the same shared uid (in this case we cannot get a Launcher3 pid if the uid is 1000 (Settings)).

In case of 1, as said before, getRunningAppProcesses works for you only because how information of getPackagesForUid and getRunningAppProcesses is returned. In this cause, Launcher3 is listed before Settings as process, because is started before than Settings.
In case of 3, the problem should be in ForegroundUtils.java not updating the foreground app properly.
Anyway, AFAIK, only case 1 could match your use case, and in that case not sure what is the best thing to do. We can check all package names to see if there is one that isn't beam disabled, but i think we can also false matches to.


Also another flaw about using getRunningAppProcesses is that RunningAppProcessInfo.proccessName not always return the packageName.


Please, debug the foregroundUids.get(0) and getPackagesForUid() to see in which case are your use case laying to.

Hope I properly explained you my concerns about this patch.

Patch-set: 4