Update patch set 1
Patch Set 1:

While I know that FeliCia and NFC-F are not the same, FeliCia building upon NFC-F, only the Japanese S5 devices support FeliCia and they do it with different HW (neither pn547 nor that brcm thing that is only in spr/dv) and respective different kernel config options. So I'd say that the hardware that even talks the wavelength isn't in the S5 at least.

And that CDD ref does state we should be properly flagging our ability to support the capability; jam-setting it to on seems like the *wrong* thing to do.

Patch-set: 1