Update patch set 1
Patch Set 1:

> Wait, so are you speaking in favor of this patch with this
 > statement?

Yes. If I'm reading the statement from the CDD correctly, capable devices must flag the feature; there is no requirement (it would be asinine to create one) that incapable devices flag the feature.

Default to on is ok if most devices don't puke on it, but there should be a way to opt out. It is a feature mask after all.

Patch-set: 1