Update patch set 1
Patch Set 1:

> I have no idea what the context here is, your commit message
 > doesn't provide enough information on what you're attempting to do.

I'm trying to get HCE working on my d2, pn544 chip, and when I added the HCE feature XML, I got JNI crashes causing constant Nfc Service crashes that I'm trying to resolve. I removed the JNI dependencies in order to fix the crashing, and I'm now trying to hardcode the values that were supposed to be retrieved from the JNI in order to get HCE working. I have confirmed the HCE works flawlessly on stock, so it's a matter of trying to connect the missing pieces to get it working.

Patch-set: 1