Update patch set 2
Patch Set 2:

If patch 1 works and it doesn't cause other issues can't we can get that merged and if a better version needs to be done work on that and merge it later? This had been an issue since the release of los and people really want it fixed.

Patch-set: 2