nxp: Add support for ISO DEP extended length
This is basically inspired on the NCI implementation merged upstream in commit 5d5dd8a.
The difference is that, instead of reading the config from an NFC configuration file,
it checks whether the feature is supported based on a system property.
The reason is that NXP's old ass stack doesn't allow reading configs from files.

Change-Id: I30083cdd55de42b7f452e4137e18f66d5f5360d4